Add a universal 'post' tag
This tag will be used for all kinds of serialized content on the site.
It will be useful for making an RSS feed later on. Both daily log
entries and more long-form posts will recieve this tag, but project
pages will not, because they are meant to be living documents, not
'posts'.