Add root-project tag
This tag will be used to organize projects that have no parent project,
for later navigation generation once I have more pages.

This is motivated by CSS Breakdown, which is a project I want in my main
projects collection, but it is still a sub-project of my blog project
page.