Refactor log folder structure and metadata
The calculated-data title of the log entries was interfering with the
log index page, with no apparent way to override a calculated data field
from the index.md file.

Previously, this was solved by putting the calculated title in the
pre-batch folder. Now that there are multiple subfolders to organize
entries, I have put all of those subfolders into an entry folder that
sits in the log folder.